Fix crash when eating bonus from right side.
The game registers Pacman eating it more than once, and so repeatedly `assign`s the `ScoreComponent`. This is illegal as we cannot assign a component that is already assigned to the GameObject.

This crash also hides the oversight that the fruit is still edible when the popup appears.

Eating from the left strangely works normally.

PS: Ghosts don't suffer from the "repeated eating" issue because the game freezes until the popup disappears, unlike the bonus popup. The ghosts would also immediately change to the `dead` state, which is specially checked by the eating logic.